requirements.txt
absl-py==2.0.0
accelerate==0.20.3
anyio==3.5.0
appdirs==1.4.4
argon2-cffi==20.1.0
attrs==22.1.0
audioread==3.0.1
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==4.1.0
Brotli==1.0.9
cachetools==5.3.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cycler==0.11.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
docker-pycreds==0.4.0
einops==0.6.1
entrypoints==0.4
fastjsonschema==2.16.2
filelock==3.12.2
flake8==5.0.4
flake8-comprehensions==3.13.0
flatbuffers==23.5.26
fonttools==4.38.0
fsspec==2023.1.0
fvcore==0.1.5.post20210915
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.25.2
google-auth-oauthlib==0.4.6
grpcio==1.60.0
huggingface-hub==0.16.4
idna==3.4
imageio==2.5.0
imageio-ffmpeg==0.3.0
importlib-metadata==6.7.0
importlib-resources==5.2.0
iopath==0.1.9
ipykernel==6.15.2
ipython==7.31.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.6
jsonschema==4.17.3
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.39.1
Markdown==3.4.4
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mediapipe==0.9.0.1
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mobile-sam==1.0
msgpack==1.0.5
multiprocess==0.70.15
nbclassic==0.5.2
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.6.3
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numpy==1.21.5
oauthlib==3.2.2
olefile==0.46
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
p-tqdm==1.4.0
packaging==23.2
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathos==0.3.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==4.0.0
pooch==1.8.0
portalocker==1.4.0
pox==0.3.3
ppft==1.7.6.7
prometheus-client==0.14.1
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.1
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
Pygments==2.11.2
pyparsing==3.1.1
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.4.0
QtPy==2.2.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.1
scikit-image==0.18.0
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.0
sentry-sdk==1.39.1
setproctitle==1.3.3
setuptools==65.6.3
six==1.16.0
smmap==5.0.1
sniffio==1.2.0
soundfile==0.12.1
soupsieve==2.3.2.post1
soxr==0.3.7
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.3.0
terminado==0.17.1
testpath==0.6.0
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.9.12
tomli==2.0.1
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tornado==6.2
tqdm==4.66.1
traitlets==5.7.1
typing_extensions==4.7.1
urllib3==2.0.4
wandb==0.16.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==2.2.3
wheel==0.38.4
widgetsnbextension==3.5.2
yacs==0.1.8
zipp==3.15.0
